refactor TemplateDetail into TemplatePage & TemplateDetails components
